chore: add deno config
