+++ index.html
@@ -95,6 +95,6 @@
       </div>
     </div>
   </div>
-  <script src="./snake.js"></script>
+<script type="module" src="./src/snake.js"></script>
 </body>
 </html>
